Correct EGL libraries and glvnd loader script.
libEGL.so.XXX.YY is the old EGL interface. If it's present, libEGL.so
and libEGL.so.1 should both link to libEGL.so.XXX.YY.

libEGL.so.1 is the new EGL interface, and only requires libEGL.so to
link to it.

This commit removes libEGL.so.XXX.YY from the package. It also correctly
includes 10_nvidia.json at /usr/share/glvnd/egl_vendor.d so that the EGL
loader is aware of the NV-provided EGL vendor library.